Manage tables and searchable fields

+++ app/controllers/manuscripts_controller.rb
@@ -49,9 +49,9 @@ class ManuscriptsController < ManageModelsController
       entries_count: obj.entries_count,
       reviewed: obj.reviewed,
       created_by: obj.created_by.present? ? obj.created_by.username : "(none)",
-      created_at: obj.created_at.present? ? obj.created_at.to_formatted_s(:short) : "",
+      created_at: obj.created_at.present? ? obj.created_at.to_formatted_s(:long) : "",
       updated_by: obj.updated_by.present? ? obj.updated_by.username : "(none)",
-      updated_at: obj.updated_at.present? ? obj.updated_at.to_formatted_s(:short) : ""
+      updated_at: obj.updated_at.present? ? obj.updated_at.to_formatted_s(:long) : ""
     }
   end
