+++ app/controllers/profiles_controller.rb
@@ -8,7 +8,7 @@ class ProfilesController < ApplicationController
       username: @user.username,
       fullname: @user.fullname || "Not set",
       email: @user.email_is_public ? @user.email : "This user's email is not publicly available.",
-      last_seen: @user.updated_at ? @user.updated_at.to_formatted_s(:short) : "",
+      last_seen: @user.updated_at ? @user.updated_at.to_formatted_s(:long) : "",
       role: @user.role,
       biography: @user.bio || "This user has chosen not to share any biographical details.",
       institution: @user.institutional_affiliation  || "Unaffiliated",
